+++ CHANGELOG.md
@@ -11,6 +11,7 @@ CHANGELOG
 - retain heatmap replay close-button after completion
 - add program configuration path to database info box in statistics
 - add journaling file size to database file size in statistics info box
+- make heatmap display settings configurable (issue #32)
 - fix logging Numpad operator keys in Linux (issue #30)
 - fix SQLite-related warnings in Python 3.12+ (issue #30)
 - use log scale in keyboard heatmaps to avoid superfrequent keys fading others (issue #32)